+++ src/main/java/tads/ufrn/apigestao/domain/dto/collector/CollectorCommissionDTO.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.YearMonth;
@@ -16,5 +17,5 @@ public class CollectorCommissionDTO {
     private String collectorName;
     private LocalDate startDate;
     private LocalDate endDate;
-    private Double commission;
+    private BigDecimal commission;
 }
