+++ src/main/java/tads/ufrn/apigestao/domain/dto/sale/ApprovePreSaleDTO.java
@@ -7,6 +7,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import tads.ufrn.apigestao.enums.PaymentType;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -14,7 +16,7 @@ public class ApprovePreSaleDTO {
     private Long inspectorId;
     private PaymentType paymentMethod;
     private int installments;
-    private Double cashPaid;
+    private BigDecimal cashPaid;
     private Double latitude;
     private Double longitude;
 }
