+++ src/main/java/tads/ufrn/apigestao/domain/dto/seller/SellerCommissionDTO.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -16,5 +17,5 @@ public class SellerCommissionDTO {
     private String sellerName;
     private LocalDate startDate;
     private LocalDate endDate;
-    private Double commission;
+    private BigDecimal commission;
 }
